requirements.txt
absl-py==2.3.1
accelerate==1.5.2
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.3.2
aiosqlite==0.22.1
albucore==0.0.24
albumentations==2.0.8
alembic==1.15.1
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.1
APScheduler==3.11.2
asteroid-filterbanks==0.4.0
asyncua==1.1.8
attrs==25.3.0
av==14.2.0
babel==2.17.0
backoff==2.2.1
basicsr==1.4.2
beautifulsoup4==4.14.3
bitsandbytes==0.47.0
blis==1.3.0
boto3==1.41.5
botocore==1.41.5
bs4==0.0.2
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.8.2
catalogue==2.0.10
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.1
cryptography==46.0.3
cssselect2==0.8.0
csvw==3.5.1
ctranslate2==4.5.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
easyocr==1.7.2
einops==0.8.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
facexlib==0.3.0
fastapi==0.115.14
faster-whisper==1.1.1
filelock==3.17.0
filterpy==1.4.5
flatbuffers==25.12.19
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
google-crc32c==1.8.0
greenlet==3.1.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
ifaddr==0.2.0
ImageIO==2.37.2
inference==0.63.5
inference-exp==0.17.3
iniconfig==2.3.0
insanely-fast-whisper==0.0.15
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.14.2
llvmlite==0.46.0
lmdb==1.7.5
lxml==6.0.2
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.31
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.2.0
murmurhash==1.0.12
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
num2words==0.5.14
numba==0.63.1
numpy==2.2.6
nvidia-ml-py==12.575.51
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.21.0
onvif-zeep-async==2.0.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
optuna==4.2.1
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pefile==2023.2.7
peft==0.17.1
phonemizer==3.3.0
pillow==11.1.0
platformdirs==4.5.1
pluggy==1.6.0
preshed==3.0.9
primePy==1.3
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.24.0
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pybase64==1.0.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.7.1
pydantic_core==2.33.1
pydot==2.0.0
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyJWT==2.10.1
pylibsrtp==1.0.0
pylogix==1.0.5
pymodbus==3.8.3
pyOpenSSL==25.3.0
pyparsing==3.2.2
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pytest==8.4.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-dotenv==1.0.1
pytorch-lightning==2.6.0
pytorch-metric-learning==2.8.1
pytorch-msssim==1.0.0
pytz==2025.1
pyvips==2.2.3
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.14.3
rdflib==7.1.3
realesrgan==0.3.0
redis==5.0.8
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==3.0.1
requests-toolbelt==1.0.0
rf-clip==1.1
rf_groundingdino==0.3.0
rfc3986==1.5.0
ric==0.0.4
rich==14.2.0
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.15.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
segmentation_models_pytorch==0.5.0
segments==2.3.0
semver==3.0.4
sentence-transformers==4.1.0
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
shiboken6==6.9.1
simple-pid==2.0.1
simsimd==6.5.12
six==1.17.0
slack_sdk==3.33.5
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.2
SQLAlchemy==2.0.39
srsly==2.5.1
srt==3.5.3
starlette==0.46.2
stringzilla==4.6.0
structlog==24.4.0
supervision==0.27.0
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.12.20
timm==1.0.24
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.22.2
torch==2.7.1+cu118
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1+cu118
torchmetrics==1.8.2
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.57.5
twilio==9.3.8
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.40.0
vosk==0.3.45
wait_for2==0.3.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
wrapt==1.17.2
yapf==0.43.0
yarl==1.18.3
zeep==4.3.2
zxing-cpp==2.2.0
